Place Pawn as FireHand
Put the FireHand mesh as the main Pawn Mesh.

As of now, it doesn't appear we can scale it, although we *can* move the
camera to put it on the left/right of it. Haven't done it right now
because to make it look well aligned, we'll need to attack the same
multiplication code as before.

Animation set code could be difficult - leaving to next build. Basic
ideas is that now we need to bind the animations for shooting as such to
these Animations the FireHand has. Not impossible, but difficult.
Possibly.