Hooks added for Animation fixes
While not implemented, the code for a possible animation fix is
commented out and put into classes it might be prudent to have it at.

Which is to say, I really don't want to have to subclass the
SkeletalMeshComponent class, so I'm hoping to see what would be the
easiest solution to moving the hand over.